Обновит взаимодействие с хранилищем в компонентах GameScreen и WinScreen
