+++ src/store/root-reducer.js
@@ -1,33 +1,13 @@
-import {extend} from "../utils";
-import {ActionType} from "./action";
-import questions from "../mocks/questions";
+import {combineReducers} from "redux";
+import {gameProcess} from "./game-process/game-process";
+import {gameData} from "./game-data/game-data";
 
-const initialState = {
-  mistakes: 0,
-  step: 0,
-  questions,
+export const NameSpace = {
+  DATA: `DATA`,
+  GAME: `GAME`,
 };
 
-const reducer = (state = initialState, action) => {
-  switch (action.type) {
-    case ActionType.INCREMENT_STEP:
-      let nextStep = state.step + action.payload;
-
-      return extend(state, {
-        step: nextStep,
-      });
-
-    case ActionType.INCREMENT_MISTAKES:
-      return extend(state, {
-        mistakes: state.mistakes + action.payload,
-      });
-
-    case ActionType.RESET_GAME:
-      return extend({}, initialState);
-  }
-
-  return state;
-};
-
-
-export {reducer};
+export default combineReducers({
+  [NameSpace.DATA]: gameData,
+  [NameSpace.GAME]: gameProcess,
+});
